requirements.txt
psycopg2
PyJWT
djangorestframework
djangorestframework-simplejwt
Django
django-cors-headers
whitenoise
setuptools
asgiref
dj-database-url
gunicorn
asgiref
pillow
django-jazzmin
urllib3
openpyxl
pandas
geopy
requests
rsa
python-decouple
django-mptt
cloudinary
qrcode[pil]
